Reorganise component folders
